Refactoring of parameters
