+++ src/CodeReview.Evaluator/Commands/IExtractMetadataCommand.cs
@@ -1,6 +1,7 @@
 using System.Threading.Tasks;
+using GodelTech.CodeReview.Evaluator.Options;
 
-namespace GodelTech.CodeReview.Evaluator.Options
+namespace GodelTech.CodeReview.Evaluator.Commands
 {
     public interface IExtractMetadataCommand
     {
