--- src/CodeReview.Evaluator/Options/ExportDbOptions.cs
@@ -1,9 +1,0 @@
-﻿using CommandLine;
-
-namespace GodelTech.CodeReview.Evaluator.Options
-{
-    [Verb("export-db", HelpText = "Create issue summary using provided manifest.")]
-    public class ExportDbOptions : IssueProcessingOptionsBase
-    {
-    }
-}
